requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==2.3.7
email-validator==2.1.0
python-dotenv==1.0.0
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-Admin==1.6.1
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.1
Flask-Script==2.0.6
Flask-Session==0.5.0
Flask-Testing==0.8.1
Flask-Babel==3.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.0
Flask-SocketIO==5.3.5
Flask-User==1.0.2.2
WTForms==3.0.1
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
python-dateutil==2.8.2
requests==2.31.0
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
isort==5.12.0
